+++ src/main/java/org/mskcc/kickoff/velox/VeloxRequestProxy.java
@@ -500,7 +500,7 @@ public class VeloxRequestProxy implements RequestProxy {
                     if (sampleLevelQcStatus != QcStatus.PASSED)
                         run.setBadRun(true);
 
-                    int totalReads = getTotalReads(sampleQc);
+                    long totalReads = getTotalReads(sampleQc);
                     run.setNumberOfReads(totalReads);
                     sample.setNumberOfReads(sample.getNumberOfReads() + totalReads);
 
@@ -523,12 +523,13 @@ public class VeloxRequestProxy implements RequestProxy {
         }
     }
 
-    private int getTotalReads(DataRecord sampleQc) throws NotFound, RemoteException {
-        int totalReads = 0;
+    private long getTotalReads(DataRecord sampleQc) throws NotFound, RemoteException {
+        long totalReads = 0;
         try {
-            totalReads = (int) (long) sampleQc.getLongVal(VeloxConstants.TOTAL_READS, user);
+            totalReads = sampleQc.getLongVal(VeloxConstants.TOTAL_READS, user);
         } catch (NullPointerException skipped) {
         }
+
         return totalReads;
     }
